Rename and move Problem APIs; Fix GetProblems
